Remove 'course' entry in ManageCoursePage.propTypes
Details: In the original code from Pluralsight the course property on the ManageCoursePage is used ONLY to set the initial state of the course state attribute to the value of the 'course' property. For clarity, I named the property 'initialCourse' instead of 'course', however I originally still had a 'course' property in ManageCoursePage.propTypes because that was how it was written in the code on Pluralsight. That was causing a validation error in my code, because since my property is named 'initialCourse' instead of 'course', there was no 'course' property being set in ManageCoursePage. This commit fixes that validation error by removing 'course' from ManageCoursePage.propTypes. The 'initialCourse' property is still in ManageCoursePage.propTypes. The references to 'course' that remain in the ManageCoursePage code are all references to the 'course' state attribute, and NOT to any 'course' property.